+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fixed an issue where the Mythic+ timer would not properly count all minions
 - Fixed a bug that was causing buttons to appear on the chat frame
 - Fixed incorrect tooltip text in the Mythic+ data panel text
+- Tooltip text now properly uses the appearance's primary color
 
 #### New
 
@@ -14,4 +15,5 @@
 - Introduced a setup wizard for onboarding
 - Re-skinned buttons that are available on notifications to the in-house skin
 - Revamp of global Appearances, including the ability to reset, global font, status texture, mute sound, class icon picker, etc.
-- Added a frame to the keystone frame which provides buttons to start a ready check, start a pull timer, and enable or disable automatically starting the key after a pull timer.
+- Added a frame to the keystone frame which provides buttons to start a ready check, start a pull timer, and enable or disable automatically starting the key after a pull timer
+- Simplified configuration down to more easy to understand sections, brought together similar configurations that were previously buried into their module's configuration